refactor: 입력, 출력을 담당하는 클래스를 io 패키지로 클래스 이동
